Moved TensorBoardLogger and tensorboard_intro into a new file.
